// src/payload-types.ts
export interface Media {
  id: number;
  alt?: string | null;
  caption?: string | null;
  prefix?: string | null;
  _objectKey?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "posts".
 */
export interface Post {
  id: number;
  title: string;
  /**
   * Permanent URL: /post/your-slug. Keep the slug unchanged after publishing.
   */
  slug: string;
  excerpt: string;
  coverImage: number | Media;
  date: string;
  readTime: string;
  /**
   * Markdown supported: headings, links, lists and images.
   */
  content: string;
  tags?:
    | {
        tag: string;
        id?: string | null;
      }[]
    | null;
  publishedAt?: string | null;
  updatedAt: string;
  createdAt: string;
  _status?: ('draft' | 'published') | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "recommendations".
 */
export interface Recommendation {
  id: number;
  name: string;
  description: string;
  type: 'Vörumerki' | 'Búð' | 'Vara';
  image: number | Media;
  link: string;
  publishedAt?: string | null;
  updatedAt: string;
  createdAt: string;
  _status?: ('draft' | 'published') | null;
}